FOGL-743 - added skeleton for sensortag device service
