+++ python/foglamp/plugins/south/sensortag/sensortag.py
@@ -9,12 +9,17 @@ import copy
 from datetime import datetime, timezone
 import uuid
 import pexpect
-import sys
 import json
 import time
 import datetime
-from foglamp.common import logger
+import sys
+import asyncio
+from aiohttp import web
 from foglamp.services.south import exceptions
+from foglamp.common import logger
+from foglamp.plugins.south.sensortag.sensortag_poll import *
+from foglamp.services.south.ingest import Ingest
+
 
 __author__ = "Amarendra K Sinha"
 __copyright__ = "Copyright (c) 2017 OSIsoft, LLC"
@@ -39,102 +44,10 @@ _DEFAULT_CONFIG = {
     }
 }
 
-characteristics = {
-    'temperature': {
-        'data': {
-            'uuid': 'f000aa01-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        },
-        'notification': {
-            'uuid': '2902',
-            'handle': '0x0000'
-        },
-        'configuration': {
-            'uuid': 'f000aa02-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        },
-        'period': {
-            'uuid': 'f000aa03-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        }
-    },
-    'movement': {
-        'data': {
-            'uuid': 'f000aa81-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        },
-        'notification': {
-            'uuid': '2902',
-            'handle': '0x0000'
-        },
-        'configuration': {
-            'uuid': 'f000aa82-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        },
-        'period': {
-            'uuid': 'f000aa83-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        }
-    },
-    'humidity': {
-        'data': {
-            'uuid': 'f000aa21-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        },
-        'notification': {
-            'uuid': '2902',
-            'handle': '0x0000'
-        },
-        'configuration': {
-            'uuid': 'f000aa22-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        },
-        'period': {
-            'uuid': 'f000aa23-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        }
-    },
-    'pressure': {
-        'data': {
-            'uuid': 'f000aa41-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        },
-        'notification': {
-            'uuid': '2902',
-            'handle': '0x0000'
-        },
-        'configuration': {
-            'uuid': 'f000aa42-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        },
-        'period': {
-            'uuid': 'f000aa43-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        }
-    },
-    'luminance': {
-        'data': {
-            'uuid': 'f000aa71-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        },
-        'notification': {
-            'uuid': '2902',
-            'handle': '0x0000'
-        },
-        'configuration': {
-            'uuid': 'f000aa72-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        },
-        'period': {
-            'uuid': 'f000aa73-0451-4000-b000-000000000000',
-            'handle': '0x0000'
-        }
-    }
-}
-
 # TODO: Implement logging
 _LOGGER = logger.setup(__name__)
 
+sensortag_characteristics = characteristics
 
 def plugin_info():
     """ Returns information about the plugin.
@@ -164,20 +77,20 @@ def plugin_init(config):
         handle: JSON object to be used in future calls to the plugin
     Raises:
     """
-    global characteristics
+    global sensortag_characteristics
 
     bluetooth_adr = config['bluetooth_adr']
     tag = SensorTag(bluetooth_adr)
 
     # The GATT table can change for different firmware revisions, so it is important to do a proper characteristic
     # discovery rather than hard-coding the attribute handles.
-    for char in characteristics.keys():
+    for char in sensortag_characteristics.keys():
         for type in ['data', 'configuration', 'period']:
-            handle = tag.get_char_handle(characteristics[char][type]['uuid'])
-            characteristics[char][type]['handle'] = handle
+            handle = tag.get_char_handle(sensortag_characteristics[char][type]['uuid'])
+            sensortag_characteristics[char][type]['handle'] = handle
 
     data = copy.deepcopy(config)
-    data['characteristics'] = characteristics
+    data['characteristics'] = sensortag_characteristics
     return data
 
 
@@ -298,284 +211,67 @@ def plugin_shutdown(handle):
     pass
 
 
+# Pending: complete below class implementation
 # TODO: Implement FOGL-701 (implement AuditLogger which logs to DB and can be used by all ) for this class
-class SensorTag(object):
-    """Handles polling of readings from SensorTag
-        Cloned from https://github.com/OrestisEv/SensorTag-Pi3
-    """
-    reading_iterations = 1  # number of iterations to read data from the TAG
+class SensorTagIngest(object):
+    """Handles incoming sensor readings from HTTP Listener"""
 
-    def __init__(self, bluetooth_adr):
-        self.con = pexpect.spawn('gatttool -b ' + bluetooth_adr + ' --interactive')
-        self.con.expect('\[LE\]>', timeout=600)
-        print("INFO: Preparing to connect. Hold on a second...If nothing happens please press the power button...")
-        self.con.sendline('connect')
-        # test for success of connect
-        self.con.expect('.*Connection successful.*\[LE\]>')
-        print("INFO: Connection Successful!")
-        self.cb = {}
+    @staticmethod
+    async def render_post(request):
+        """Store sensortag readings to FogLAMP
+        """
 
-    def get_char_handle(self, uuid):
-        timeout = 3
-        max_time = time.time() + timeout
-        rval = '0x0000'
-        while time.time() < max_time:
-            try:
-                cmd = 'char-read-uuid %s' % uuid
-                self.con.sendline(cmd)
-                # TODO: Devise a better method for all below
-                self.con.expect('.*handle:.* \r', timeout=3)
-                reading = self.con.after
-                line = reading.decode().split('handle: ')[1]
-                # print(uuid, reading, "\n") #line.split()[0])
-                rval = line.split()[0]
-            except Exception as ex:
-                time.sleep(.5)
+        increment_discarded_counter = False
+
+        # TODO: Decide upon the correct format of message
+        message = {'result': 'success'}
+        code = web.HTTPOk.status_code
+
+        try:
+            if not Ingest.is_available():
+                increment_discarded_counter = True
+                message = {'busy': True}
             else:
-                break
-        return rval
+                payload = await request.json()
 
-    def char_write_cmd(self, handle, value):
-        # The 0%x for value is VERY naughty!  Fix this!
-        cmd = 'char-write-cmd %s %s' % (handle, value)
-        self.con.sendline(cmd)
-        # delay for 1 second so that Tag can enable registers
-        time.sleep(1)
+                if not isinstance(payload, dict):
+                    raise ValueError('Payload must be a dictionary')
 
-    def char_read_hnd(self, handle, sensortype):
-        # send the hex value to the Tag
-        self.con.sendline('char-read-hnd %s' % handle)
+                asset = payload.get('asset')
+                timestamp = payload.get('timestamp')
+                key = payload.get('key')
 
-        self.con.expect('.*descriptor:.* \r')
-        reading = self.con.after
-        print("DEBUGGING: Reading from Tag... %s \n" % reading) #print(the outcome as it comes while reading the Tag
-        rval = reading.split() #splitting the reading based on the spaces
-        # print("DEBUGGING: rval" + str(rval)+"\n")
+                # readings and sensor_readings are optional
+                try:
+                    readings = payload.get('readings')
+                except KeyError:
+                    readings = payload.get('sensor_values')  # sensor_values is deprecated
 
-        if sensortype in ['temperature']:
-            # The raw data value read from this sensor are two unsigned 16 bit values
-            raw_measurement = rval[-4] + rval[-3] + rval[-2] + rval[-1]
-        elif sensortype in ['movement']:
-            # TODO: Pending implementation
-            pass
-        elif sensortype in ['humidity']:
-            # The raw data value read from this sensor are two unsigned 16 bit values
-            raw_measurement = rval[-4] + rval[-3] + rval[-2] + rval[-1]
-        elif sensortype in ['pressure']:
-            # The data from the pressure sensor consists of two 24-bit unsigned integers
-            raw_measurement = rval[-6] + rval[-5] + rval[-4] + rval[-3] + rval[-2] + rval[-1]
-        elif sensortype in ['luminance']:
-            # The data from the optical sensor consists of a single 16-bit unsigned integer
-            raw_measurement = rval[-2] + rval[-1]
-        else:
-            raw_measurement = 0
+                if not isinstance(readings, dict):
+                    raise ValueError('readings must be a dictionary')
 
-        print(sensortype, raw_measurement, "\n")
-        return raw_measurement
+                await Ingest.add_readings(asset=asset, timestamp=timestamp, key=key, readings=readings)
+        except (ValueError, TypeError) as e:
+            increment_discarded_counter = True
+            code = web.HTTPBadRequest.status_code
+            message = {'error': str(e)}
+            _LOGGER.exception(str(e))
+        except Exception as e:
+            increment_discarded_counter = True
+            code = web.HTTPInternalServerError.status_code
+            message = {'error': str(e)}
+            _LOGGER.exception(str(e))
 
-    @staticmethod
-    def hexTemp2C(raw_tempr):
-        """
-        Conversion method at http://processors.wiki.ti.com/index.php/CC2650_SensorTag_User's_Guide#Gatt_Server
-        The raw data value read from this sensor are two unsigned 16 bit values, one for die (ambience) temperature and
-        one for object temperature. To convert to temperature in degrees (Celsius), use the algorithm below:
-        void sensorTmp007Convert(uint16_t rawAmbTemp, uint16_t rawObjTemp, float *tAmb, float *tObj)
-        {
-          const float SCALE_LSB = 0.03125;
-          float t;
-          int it;
+        if increment_discarded_counter:
+            Ingest.increment_discarded_readings()
 
-          it = (int)((rawObjTemp) >> 2);
-          t = ((float)(it)) * SCALE_LSB;
-          *tObj = t;
+        # expect keys in response:
+        # (code = 2xx) result Or busy
+        # (code = 4xx, 5xx) error
+        message['status'] = code
 
-          it = (int)((rawAmbTemp) >> 2);
-          t = (float)it;
-          *tAmb = t * SCALE_LSB;
-        }
-        :param raw_tempr:
-        :return:
-        """
-        raw_temperature = raw_tempr.decode()
-        SCALE_LSB = 0.03125
+        return web.json_response(message)
 
-        # Choose object temperature (reverse bytes for little endian)
-        raw_object_temp = int('0x' + raw_temperature[2:4] + raw_temperature[0:2], 16)
-        object_temp_int = raw_object_temp >> 2
-        object_temp_celsius = float(object_temp_int) * SCALE_LSB
-
-        # Choose ambient temperature (reverse bytes for little endian)
-        raw_ambient_temp = int('0x' + raw_temperature[6:8] + raw_temperature[4:6], 16)
-        ambient_temp_int = raw_ambient_temp >> 2
-        ambient_temp_celsius = float(ambient_temp_int) * SCALE_LSB
-        ambient_temp_fahrenheit = (ambient_temp_celsius * 1.8) + 32
-
-        print("INFO: object Celsius:    %f" % object_temp_celsius)
-        print("INFO: Ambient Celsius:    %f" % ambient_temp_celsius)
-        print("Fahrenheit: %f" % ambient_temp_fahrenheit)
-
-        return object_temp_celsius, ambient_temp_celsius
-
-    @staticmethod
-    def hexMovement2Mov(raw_movement):
-        """
-        Conversion method at http://processors.wiki.ti.com/index.php/CC2650_SensorTag_User's_Guide#Gatt_Server
-
-        Gyroscope raw data make up 0-5 of the data from the movement service, in the order X, Y, Z axis. Data from
-        each axis consists of two bytes, encoded as a signed integer. For conversion from gyroscope raw data to
-        degrees/second, use the algorithm below on each of the first three 16-bit values in the incoming data, one for
-        each axis. Note that the axis data from a disabled axis will be 0, so the size of the incoming data is always
-        18 bytes. When the WOM feature is enabled, the latest measured data will be continuously transmitted.
-        float sensorMpu9250GyroConvert(int16_t data)
-        {
-          //-- calculate rotation, unit deg/s, range -250, +250
-          return (data * 1.0) / (65536 / 500);
-        }
-
-        Accelerometer raw data make up bytes 6-11 of the data from the movement service, in the order X, Y, Z axis.
-        Data from each axis consists of two bytes, encoded as a signed integer. For conversion from accelerometer raw
-        data to Gravity (G), use the algorithm below on each the three 16-bit values in the incoming data, one for each
-        axis.
-        // Accelerometer ranges
-        #define ACC_RANGE_2G      0
-        #define ACC_RANGE_4G      1
-        #define ACC_RANGE_8G      2
-        #define ACC_RANGE_16G     3
-
-        float sensorMpu9250AccConvert(int16_t rawData)
-        {
-          float v;
-
-          switch (accRange)
-          {
-          case ACC_RANGE_2G:
-            //-- calculate acceleration, unit G, range -2, +2
-            v = (rawData * 1.0) / (32768/2);
-            break;
-
-          case ACC_RANGE_4G:
-            //-- calculate acceleration, unit G, range -4, +4
-            v = (rawData * 1.0) / (32768/4);
-            break;
-
-          case ACC_RANGE_8G:
-            //-- calculate acceleration, unit G, range -8, +8
-            v = (rawData * 1.0) / (32768/8);
-            break;
-
-          case ACC_RANGE_16G:
-            //-- calculate acceleration, unit G, range -16, +16
-            v = (rawData * 1.0) / (32768/16);
-            break;
-          }
-
-          return v;
-        }
-
-        Magnetometer raw data make up bytes 12-17 of the data from the movement service, in the order X, Y, Z axis.
-        Data from each axis consists of two bytes, encoded as a signed integer. The conversion is done in the
-        SensorTag firmware so there is no calculation involved apart from changing the integer to a float if required.
-        The measurement unit is uT (micro Tesla).
-        float sensorMpu9250MagConvert(int16_t data)
-        {
-          //-- calculate magnetism, unit uT, range +-4900
-          return 1.0 * data;
-        }
-        :param raw_movement:
-        :return:
-        """
-        # To be implemented
-        pass
-
-    @staticmethod
-    def hexHum2RelHum(raw_humd):
-        """
-        Conversion method at http://processors.wiki.ti.com/index.php/CC2650_SensorTag_User's_Guide#Gatt_Server
-
-        The data from the humidity sensor consists of two 16-bit unsigned integers: the temperature in bytes 0 and 1,
-        and the pressure in bytes 2 and 3. Conversion to temperature and relative humidity is done as shown below.
-        The temperature unit is degrees Celsius (°C), the humidity unit is relative humidity (%RH).
-        void sensorHdc1000Convert(uint16_t rawTemp, uint16_t rawHum,
-                        float *temp, float *hum)
-        {
-          //-- calculate temperature [°C]
-          *temp = ((double)(int16_t)rawTemp / 65536)*165 - 40;
-
-          //-- calculate relative humidity [%RH]
-          rawHum &= ~0x0003; // remove status bits
-          *hum = ((double)rawHum / 65536)*100;
-        }
-        :param raw_humd:
-        :return:
-        """
-        interim_value = raw_humd.decode()
-        raw_temperature = int('0x'+interim_value[2:4]+interim_value[0:2], 16)
-        temperature = ((raw_temperature / 65536) * 165) - 40
-        raw_humidity = int('0x'+interim_value[6:8]+interim_value[4:6], 16)
-        raw_humidity &= -0x0003
-        humidity = float((raw_humidity)) / 65536 * 100
-        print("INFO: tempr:    %f" % temperature)
-        print("INFO: humidity:    %f" % humidity)
-        return humidity
-
-    @staticmethod
-    def hexPress2Press(raw_pr):
-        """
-        Conversion method at http://processors.wiki.ti.com/index.php/CC2650_SensorTag_User's_Guide#Gatt_Server
-
-        The data from the pressure sensor consists of two 24-bit unsigned integers: the temperature in bytes 0-2,
-        the pressure in bytes 3-5. The conversion and adjustment calculations is done in firmware, so the
-        application in effect only has to divide the incoming values by 100. Conversion to temperature and
-        pressure is done as shown below. The temperature unit is degrees Celsius, the pressure in hectopascal (hPa).
-        float sensorBmp280Convert(uint32_t rawValue)
-        {
-          return rawValue / 100.0f;
-        }
-        :param raw_pr:
-        :return:
-        """
-        interim_value = raw_pr.decode()
-        raw_pressure = int('0x'+interim_value[10:12]+interim_value[8:10]+interim_value[6:8], 16)
-        pressure = float(raw_pressure) / 100.0
-        print("INFO: pressure:    %f" % pressure)
-        return pressure
-
-    @staticmethod
-    def hexLum2Lux(raw_lumn):
-        """
-        Conversion method at http://processors.wiki.ti.com/index.php/CC2650_SensorTag_User's_Guide#Gatt_Server
-
-        The data from the optical sensor consists of a single 16-bit unsigned integer. Conversion to l
-        ight intensity (LUX) is shown below.
-        float sensorOpt3001Convert(uint16_t rawData)
-        {
-            uint16_t e, m;
-
-            m = rawData & 0x0FFF;
-            e = (rawData & 0xF000) >> 12;
-
-            /** e on 4 bits stored in a 16 bit unsigned => it can store 2 << (e - 1) with e < 16 */
-            e = (e == 0) ? 1 : 2 << (e - 1);
-
-            return m * (0.01 * e);
-        }
-        :param raw_lumn:
-        :return:
-        """
-        interim_value = raw_lumn.decode()
-        # Reverse LSB and MSB
-        raw_luminance = int('0x'+interim_value[2:4]+interim_value[0:2], 16)
-        m = "0FFF"
-        e = "F000"
-        m = int(m, 16)
-        exp = int(e, 16)
-        m = (raw_luminance & m)
-        exp = (raw_luminance & exp) >> 12
-        exp = 1 if exp == 0 else 2
-        exp = exp << (exp -1)
-        luminance = (m * (0.01 * exp))
-        print("INFO: luminance:    %f" % luminance)
-        return luminance
 
 if __name__ == "__main__":
     # To run: python3 python/foglamp/plugins/south/sensortag/sensortag.py B0:91:22:EA:79:04
